+++ packages/animations/CHANGELOG.md
@@ -1,3 +1,18 @@
+## [1.0.0+5] - February 21, 2020
+
+* Fix override analyzer ignore placement.
+
+
+## [1.0.0+4] - February 21, 2020
+
+* Fix a typo in the changelog dates
+* Revert use of modern Material text style nomenclature in the example app
+  to be compatible with Flutter's `stable` branch for the time being.
+* Add override analyzer ignore in modal.dart for reverseTransitionDuration
+  until Flutter's stable branch contains
+  https://github.com/flutter/flutter/pull/48274.
+
+
 ## [1.0.0+3] - February 18, 2020
 
 * Update README.md to better describe Material motion